+++ apps/ledger-live-desktop/src/renderer/screens/market/MarketCoin/components/MarketInfo.tsx
