+++ apps/ledger-live-desktop/src/renderer/screens/market/MarketList/MarketItemChart.tsx
