+++ apps/ledger-live-desktop/src/renderer/screens/market/hooks/useMarket.ts
@@ -1,18 +1,29 @@
+import { useFetchCurrencyFrom } from "@ledgerhq/live-common/exchange/swap/hooks/index";
 import { MarketState } from "@ledgerhq/live-common/market/types";
 import { rangeDataTable } from "@ledgerhq/live-common/market/utils/rangeDataTable";
-import { useMarketDataProvider } from "@ledgerhq/live-common/market/v2/useMarketDataProvider";
+import {
+  useMarketDataProvider,
+  useMarketData as useMarketDataHook,
+} from "@ledgerhq/live-common/market/v2/useMarketDataProvider";
 import { useCallback, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { useDispatch, useSelector } from "react-redux";
 import { setMarketOptions } from "~/renderer/actions/market";
+import { addStarredMarketCoins, removeStarredMarketCoins } from "~/renderer/actions/settings";
 import { useInitSupportedCounterValues } from "~/renderer/hooks/useInitSupportedCounterValues";
 import { marketSelector } from "~/renderer/reducers/market";
-import { starredMarketCoinsSelector } from "~/renderer/reducers/settings";
+import { localeSelector, starredMarketCoinsSelector } from "~/renderer/reducers/settings";
+
+export type MarketHookResult = {};
 
 export function useMarket() {
   const { t } = useTranslation();
   const dispatch = useDispatch();
   const marketParams = useSelector(marketSelector);
+  const locale = useSelector(localeSelector);
+
+  useInitSupportedCounterValues();
+
   const { supportedCounterCurrencies } = useMarketDataProvider();
 
   const setCounterCurrency = useCallback(
@@ -38,8 +49,6 @@ export function useMarket() {
   const { range, starred = [], liveCompatible } = marketParams;
   const starredMarketCoins: string[] = useSelector(starredMarketCoinsSelector);
   const starFilterOn = starred.length > 0;
-
-  useInitSupportedCounterValues();
 
   const updateSearch = useCallback(
     (value: string) => {
@@ -78,20 +87,86 @@ export function useMarket() {
 
   const timeRangeValue = timeRanges.find(({ value }) => value === range);
 
+  const { supportedCurrencies, liveCoinsList } = useMarketDataProvider();
+
+  const onLoadNextPage = useCallback(() => {
+    dispatch(setMarketOptions({ page: (marketParams?.page || 1) + 1 }));
+  }, [dispatch, marketParams?.page]);
+
+  const { orderBy, order } = marketParams;
+
+  const marketResult = useMarketDataHook({
+    ...marketParams,
+    liveCoinsList,
+    supportedCoinsList: supportedCurrencies,
+  });
+
+  const currenciesLength = marketResult.data.length;
+  const loading = marketResult.loading;
+  const freshLoading = loading && !currenciesLength;
+
+  const { data: fromCurrencies } = useFetchCurrencyFrom();
+
+  const resetSearch = useCallback(() => refresh({ search: "" }), [refresh]);
+
+  const toggleStar = useCallback(
+    (id: string, isStarred: boolean) => {
+      if (isStarred) {
+        dispatch(removeStarredMarketCoins(id));
+      } else {
+        dispatch(addStarredMarketCoins(id));
+      }
+    },
+    [dispatch],
+  );
+
+  const toggleSortBy = useCallback(
+    (newOrderBy: string) => {
+      const isFreshSort = newOrderBy !== orderBy;
+      refresh(
+        isFreshSort
+          ? { orderBy: newOrderBy, order: "desc" }
+          : {
+              orderBy: newOrderBy,
+              order: order === "asc" ? "desc" : "asc",
+            },
+      );
+    },
+    [order, orderBy, refresh],
+  );
+
+  const isItemLoaded = useCallback(
+    (index: number) => !!marketResult.data[index],
+    [marketResult.data],
+  );
+  const itemCount = currenciesLength + 1;
+
   return {
-    timeRangeValue,
+    isItemLoaded,
+    onLoadNextPage,
     toggleLiveCompatible,
     toggleFilterByStarredAccounts,
+    toggleSortBy,
+    toggleStar,
     updateSearch,
     updateTimeRange,
-    marketParams,
+    refresh,
+    resetSearch,
     setCounterCurrency,
+    freshLoading,
     supportedCounterCurrencies,
     t,
-    refresh,
     liveCompatible,
     starFilterOn,
+    marketResult,
     starredMarketCoins,
     timeRanges,
+    marketParams,
+    timeRangeValue,
+    itemCount,
+    locale,
+    fromCurrencies,
+    loading,
+    currenciesLength,
   };
 }
